requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
asyncpg==0.30.0
httpx==0.28.1
APScheduler==3.11.0
yt-dlp==2025.10.14
anthropic==0.60.0
gspread==6.2.1
google-auth==2.40.3
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
